feat: 존재하지 않는 상품 삭제 예외 처리

+++ server/src/domain/product/product.service.ts
@@ -1,3 +1,4 @@
+import AppError from "../../errors/AppError.js";
 import Product, { ProductType } from "../../model/Product.js";
 import { ProductRepository } from "./product.repository.js";
 
@@ -17,6 +18,10 @@ class ProductService {
   }
 
   deleteProduct(id: number) {
+    const targetProduct = this.productRepository.findById(id);
+    if (!targetProduct) {
+      throw new AppError("PRODUCT_NOT_EXIST");
+    }
     this.productRepository.delete(id);
   }
